+++ packages/backend-plugin-api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/backend-plugin-api
 
+## 0.1.5-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.17.0-next.1
+  - @backstage/backend-tasks@0.4.0-next.1
+  - @backstage/config@1.0.5-next.1
+  - @backstage/plugin-permission-common@0.7.2-next.1
+
 ## 0.1.5-next.0
 
 ### Patch Changes