+++ plugins/azure-devops/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-azure-devops
 
+## 0.2.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.1
+  - @backstage/core-plugin-api@1.1.1-next.1
+  - @backstage/plugin-catalog-react@1.2.2-next.1
+  - @backstage/catalog-model@1.1.4-next.1
+  - @backstage/errors@1.1.4-next.1
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-azure-devops-common@0.3.0
+
 ## 0.2.3-next.0
 
 ### Patch Changes