+++ plugins/example-todo-list/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @internal/plugin-todo-list
 
+## 1.0.8-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.1
+  - @backstage/core-plugin-api@1.1.1-next.1
+  - @backstage/theme@0.2.16
+
 ## 1.0.8-next.0
 
 ### Patch Changes