+++ plugins/firehydrant/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-firehydrant
 
+## 0.1.29-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.1
+  - @backstage/core-plugin-api@1.1.1-next.1
+  - @backstage/plugin-catalog-react@1.2.2-next.1
+  - @backstage/theme@0.2.16
+
 ## 0.1.29-next.0
 
 ### Patch Changes