+++ plugins/kafka-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-kafka-backend
 
+## 0.2.32-next.1
+
+### Patch Changes
+
+- 0b54dbaf7d: Internal tweak to adapt to `kafkajs` 2.2.3
+- Updated dependencies
+  - @backstage/backend-common@0.17.0-next.1
+  - @backstage/config@1.0.5-next.1
+  - @backstage/catalog-model@1.1.4-next.1
+  - @backstage/errors@1.1.4-next.1
+
 ## 0.2.32-next.0
 
 ### Patch Changes